Remove Lodash and Bluebird to decrease output size.
simple-vue-validator requires both Lodash and Bluebird. According to my
Webpack dev build, Lodash@0.9.9 is 211.3k and Bluebird@3.4.7 is 173.6k.

Since both of these libraries are used minimally in
simple-vue-validator, it makes sense to replace them. This will avoid
forcing both Lodash and Bluebird onto developers who don't use them
otherwise in their project.

Bluebird:

This PR replace Bluebird with the much smaller es6-promise library.
es6-promise weighs in at 27.9k uncompressed and unminified.

es6-promise is limited in scope, which is how it can be so small. If a
feature is needed that es6-promise doesn't provide, then Bluebird can be
reconsidered. As it stands, es6-promise appears to meet the needs of
simple-vue-validator.

Lodash:

Usage of Lodash in simple-vue-validator is limited to a handful of
functions. These functions are simple enough that we can write our own
definitions, which I've done and included in `src/utils.js`.

`_.keys` was replaced with `Object.keys` which is supported by IE9+ and
all recent versions of other major browsers.

See
[here](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/keys#Browser_compatibility)
for more info on `Object.keys` browser compatibility.

`_.extend` was used once to copy the protype of `Rule`. This is a very
simple use case, so I replaced `_.extend` with a simple for loop. If
more functionality is needed, then the
[extend](https://www.npmjs.com/package/extend) NPM package might be a
good option.

See `src/utils.js` for other function definitions.

Hope this helps!